+++ data/task.txt
@@ -1,3 +1,5 @@
 [E]  &##& false &##& wdwdwdw &##& 2019-12-02T16:48 &##& 2019-12-03T23:00
 [E]  &##& false &##& fish &##& 2019-12-05T18:00 &##& 2019-12-06T18:00
 [T]  &##& true &##& yummy
+[D]  &##& false &##& done &##& 2023-02-25T19:30
+[E]  &##& false &##& homework &##& 2023-09-10T18:00 &##& 2023-09-20T18:00
