+++ src/main/java/bot/command/Command.java
@@ -1,0 +1,13 @@
+package bot.command;
+
+import bot.exception.FileErrorBotException;
+import bot.exception.IllegalExpressionBotException;
+
+import java.io.IOException;
+
+public abstract class Command {
+
+    public static final String SPACER = "____________________________________________________________";
+
+    public abstract void execute() throws IllegalExpressionBotException, FileErrorBotException, IOException;
+}
