+++ src/main/java/bot/command/DeleteCommand.java
@@ -1,9 +1,11 @@
-package command;
+package bot.command;
 
-import exception.FileErrorBotException;
-import storage.Storage;
-import task.Task;
-import task.TaskList;
+import bot.exception.FileErrorBotException;
+import bot.storage.Storage;
+import bot.task.Task;
+import bot.task.TaskList;
+
+import java.io.IOException;
 
 public class DeleteCommand extends Command {
     private final TaskList taskList;
@@ -16,7 +18,7 @@ public class DeleteCommand extends Command {
         this.task = this.taskList.get(this.idx);
     }
 
-    public void execute() throws FileErrorBotException {
+    public void execute() throws FileErrorBotException, IOException {
         this.taskList.delete(idx);
         Storage.save(this.taskList);
         System.out.println(this);
@@ -25,7 +27,7 @@ public class DeleteCommand extends Command {
     @Override
     public String toString() {
         return Command.SPACER + "\n" +
-                "Noted. I've removed this task:\n" +
+                "Noted. I've removed this bot.task:\n" +
                 this.task + "\n" +
                 "Now you have " + this.taskList.length() + " tasks in the list.\n" +
                 Command.SPACER;
