+++ src/main/java/bot/command/TodoCommand.java
@@ -1,9 +1,11 @@
-package command;
+package bot.command;
 
-import exception.FileErrorBotException;
-import task.TaskList;
-import task.Todo;
-import storage.Storage;
+import bot.exception.FileErrorBotException;
+import bot.task.TaskList;
+import bot.task.Todo;
+import bot.storage.Storage;
+
+import java.io.IOException;
 
 public class TodoCommand extends Command {
 
@@ -15,7 +17,7 @@ public class TodoCommand extends Command {
         this.todo = new Todo(taskDetail);
     }
 
-    public void execute() throws FileErrorBotException {
+    public void execute() throws FileErrorBotException, IOException {
         this.taskList.add(this.todo);
         Storage.save(this.taskList);
         System.out.println(this);
@@ -25,13 +27,13 @@ public class TodoCommand extends Command {
     public String toString() {
         if (this.taskList.length() <= 1) {
             return Command.SPACER + "\n" +
-                    "Got it. I've added this task:\n" +
+                    "Got it. I've added this bot.task:\n" +
                     this.todo + "\n" +
-                    "Now you have " + this.taskList.length() + " task in the list.\n" +
+                    "Now you have " + this.taskList.length() + " bot.task in the list.\n" +
                     Command.SPACER;
         } else {
             return Command.SPACER + "\n" +
-                    "Got it. I've added this task:\n" +
+                    "Got it. I've added this bot.task:\n" +
                     this.todo + "\n" +
                     "Now you have " + this.taskList.length() + " tasks in the list.\n" +
                     Command.SPACER;
