+++ src/main/java/bot/exception/DateTimeParseBotException.java
@@ -1,4 +1,4 @@
-package exception;
+package bot.exception;
 
 public class DateTimeParseBotException extends BotException {
     public DateTimeParseBotException(String errorMessage) {
