+++ src/main/java/bot/exception/FileErrorBotException.java
@@ -1,4 +1,4 @@
-package exception;
+package bot.exception;
 
 public class FileErrorBotException extends BotException {
     public FileErrorBotException(String errorMessage) {
