+++ src/main/java/bot/exception/IllegalExpressionBotException.java
@@ -1,4 +1,4 @@
-package exception;
+package bot.exception;
 
 public class IllegalExpressionBotException extends BotException {
     public IllegalExpressionBotException(String errorMessage) {
