+++ src/main/java/bot/exception/IncompleteBotException.java
@@ -1,4 +1,4 @@
-package exception;
+package bot.exception;
 
 public class IncompleteBotException extends BotException {
     public IncompleteBotException(String errorMessage) {
