+++ src/main/java/bot/storage/Storage.java
@@ -1,10 +1,10 @@
-package storage;
+package bot.storage;
 
-import exception.DateTimeParseBotException;
-import task.Task;
-import exception.FileErrorBotException;
-import task.TaskList;
-import parsers.StorageParser;
+import bot.exception.DateTimeParseBotException;
+import bot.task.Task;
+import bot.exception.FileErrorBotException;
+import bot.task.TaskList;
+import bot.parsers.StorageParser;
 
 import java.io.BufferedWriter;
 import java.io.File;
@@ -22,38 +22,40 @@ public class Storage {
     private static final String ABSOLUTE_FILE_PATH = Storage.WORKING_DIRECTORY + File.separator +
             Storage.FILE_RELATIVE_PATH;
 
-    public static void save(TaskList taskList) throws FileErrorBotException {
+    public static void save(TaskList taskList) throws FileErrorBotException, IOException {
         File file = new File(Storage.ABSOLUTE_FILE_PATH +
                 File.separator + Storage.FILE_NAME);
         file.getParentFile().mkdirs();
-        try {
-            BufferedWriter bw = new BufferedWriter(new FileWriter(file));
+        if (!file.exists()) {
+            file.createNewFile();
+        }
+        try (BufferedWriter bw = new BufferedWriter(new FileWriter(file))) {
             Iterator<Task> iterator = taskList.iterator();
             while(iterator.hasNext()) {
                 bw.write(iterator.next().fileWriteFormatted());
                 bw.newLine();
             }
-            bw.close();
         } catch (IOException exception) {
             throw new FileErrorBotException("File cannot be saved...");
         }
     }
 
-    public static TaskList read() throws FileErrorBotException, DateTimeParseBotException {
+    public static TaskList read() throws FileErrorBotException,
+            DateTimeParseBotException, IOException {
         File file = new File(Storage.ABSOLUTE_FILE_PATH +
                 File.separator + Storage.FILE_NAME);
         file.getParentFile().mkdirs();
+        if (!file.exists()) {
+            file.createNewFile();
+        }
         TaskList taskList = new TaskList();
-        try {
-            BufferedReader br = new BufferedReader(new FileReader(file));
+        try (BufferedReader br = new BufferedReader(new FileReader(file))) {
             while (br.ready()) {
                 taskList.add(StorageParser.parseFileInput(br.readLine()));
             }
-            br.close();
         } catch (IOException exception) {
             throw new FileErrorBotException("File cannot be saved...");
         }
         return taskList;
     }
-
 }
