+++ src/main/java/bot/task/Task.java
@@ -1,4 +1,4 @@
-package task;
+package bot.task;
 
 public abstract class Task {
     private static final String COMPLETE = "[X] ";
