--- src/main/java/command/Command.java
@@ -1,11 +1,0 @@
-package command;
-
-import exception.FileErrorBotException;
-import exception.IllegalExpressionBotException;
-
-public abstract class Command {
-
-    public static final String SPACER = "____________________________________________________________";
-
-    public abstract void execute() throws IllegalExpressionBotException, FileErrorBotException;
-}
